Move Menu button and Small hover button paths
